........
r47635 | kpfleming | 2006-11-14 18:05:44 -0600 (Tue, 14 Nov 2006) | 2 lines

silence compiler warning on 64-bit platforms (this variable is an 'int' anyway, comparing it to 'signed long' is not useful)

........


git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@47636 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ channels/misdn/isdn_lib.c
@@ -3474,7 +3474,7 @@ static int queue_l2l3(msg_t *msg)
 int manager_isdn_handler(iframe_t *frm ,msg_t *msg)
 {  
 
-	if (frm->dinfo==(signed long)0xffffffff && frm->prim==(PH_DATA|CONFIRM)) {
+	if (frm->dinfo==0xffffffff && frm->prim==(PH_DATA|CONFIRM)) {
 		cb_log(0,0,"SERIOUS BUG, dinfo == 0xffffffff, prim == PH_DATA | CONFIRM !!!!\n");
 	}
